requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.12.0
requests==2.31.0
google-generativeai>=0.3.2
droidrun>=0.3.2

websockets>=12.0
jinja2>=3.1.3
